+++ vortex-file/src/writer.rs
@@ -3,7 +3,7 @@ use vortex_array::stream::ArrayStream;
 use vortex_error::{vortex_bail, vortex_err, VortexExpect, VortexResult};
 use vortex_flatbuffers::{FlatBuffer, FlatBufferRoot, WriteFlatBuffer, WriteFlatBufferExt};
 use vortex_io::VortexWrite;
-use vortex_layout::strategies::{LayoutStrategy, VortexLayoutStrategy};
+use vortex_layout::strategy::{LayoutStrategy, VortexLayoutStrategy};
 
 use crate::footer::{FileLayout, Postscript, Segment};
 use crate::segments::writer::BufferedSegmentWriter;
