+++ vortex-layout/src/lib.rs
@@ -10,11 +10,14 @@ mod reader;
 use std::fmt::{Display, Formatter};
 
 pub use reader::*;
+pub use strategy::*;
+pub use writer::*;
 
 pub mod scan;
 pub mod segments;
-pub mod strategies;
+mod strategy;
 pub mod vtable;
+mod writer;
 
 /// The layout ID for a flat layout
 pub(crate) const FLAT_LAYOUT_ID: LayoutId = LayoutId(1);
