+++ vortex-layout/src/writer.rs
@@ -1,25 +1,10 @@
-//! This is a collection of built-in layout strategies designed to be used in conjunction with one
-//! another to develop an overall strategy.
-//!
-//! Each [`LayoutWriter`] is passed horizontal chunks of a Vortex array one-by-one, and is
-//! eventually asked to return a [`Layout`]. The writers can buffer, re-chunk, flush, or
-//! otherwise manipulate the chunks of data enabling experimentation with different strategies
-//! all while remaining independent of the read code.
-
-mod strategy;
-
-pub use strategy::*;
 use vortex_array::Array;
-use vortex_dtype::DType;
 use vortex_error::VortexResult;
 
-use crate::layouts::flat::writer::FlatLayoutWriter;
-use crate::layouts::flat::FlatLayout;
 use crate::segments::SegmentWriter;
 use crate::Layout;
 
 /// A strategy for writing chunks of an array into a layout.
-/// FIXME(ngates): move this into writer.rs
 // [layout writer]
 pub trait LayoutWriter: Send {
     fn push_chunk(&mut self, segments: &mut dyn SegmentWriter, chunk: Array) -> VortexResult<()>;
@@ -58,15 +43,3 @@ pub trait LayoutWriterExt: LayoutWriter {
 }
 
 impl<L: LayoutWriter> LayoutWriterExt for L {}
-
-/// A trait for creating new layout writers given a DType.
-pub trait LayoutStrategy: Send + Sync {
-    fn new_writer(&self, dtype: &DType) -> VortexResult<Box<dyn LayoutWriter>>;
-}
-
-/// Implement the [`LayoutStrategy`] trait for the [`FlatLayout`] for easy use.
-impl LayoutStrategy for FlatLayout {
-    fn new_writer(&self, dtype: &DType) -> VortexResult<Box<dyn LayoutWriter>> {
-        Ok(FlatLayoutWriter::new(dtype.clone(), Default::default()).boxed())
-    }
-}
